+++ Frame.h
@@ -24,12 +24,12 @@ class Frame{
     int tag;
     SDL_Surface *image;
     SDL_Rect src, dest;
-    long time;
-public:
-    Frame(string imageName, long newTime = LONG_MAX);
-    Frame(string imageName, int x, int y, int w, int h, long newTime);//for use with sprite sheet
+    Uint32 time;
+public: 
+    Frame(string imageName, Uint32 newTime = UINT32_MAX);
+    Frame(string imageName, int x, int y, int w, int h, Uint32 newTime);//for use with sprite sheet
     Frame(string fname, int column, int row);//single sprite image
-    long getTime();
+    Uint32 getTime();
     void draw(SDL_Surface *screen, int x, int y);
     SDL_Rect getRect();
     ~Frame();
