+++ Sprite.h
@@ -18,7 +18,7 @@ class Sprite{
 protected:
     Animation *animation;
     float xPos, yPos, xVel, yVel, xAcc, yAcc;
-    long last;
+    Uint32 last;
     bool isDragging;
     int offsetX, offsetY;
 public:
@@ -26,8 +26,8 @@ public:
            float newXPos = 0.0, float newYPos = 0.0,
            float newXVel = 0.0, float newYVel = 0.0,
            float newXAcc = 0.0, float newYAcc = 0.0);
-    virtual void update(long elapsed);
-    virtual void draw(SDL_Surface *screen, long elapsed );
+    virtual void update(Uint32 elapsed);
+    virtual void draw(SDL_Surface *screen, Uint32 elapsed );
     virtual SDL_Rect getRect();
     virtual bool isAlive();
     virtual void dragTo(int x, int y);
